add ponyfill exports
