+++ index.cjs
@@ -1,0 +1,7 @@
+const { URLPattern } = require("./dist/pony.cjs");
+
+module.exports = { URLPattern };
+
+if (!globalThis.URLPattern) {
+  globalThis.URLPattern = URLPattern;
+}
