--- src/patch-global.mts
@@ -1,6 +1,0 @@
-import { URLPattern } from "./url-pattern";
-export { URLPattern };
-
-if (!globalThis.URLPattern) {
-  globalThis.URLPattern = URLPattern;
-}
