+++ test/can-load-as-esm.mjs
@@ -1,17 +1,16 @@
-import "../dist/index.js";
-import  test  from "ava";
+import "urlpattern-polyfill";
+import test from "ava";
 
-const baseURL = 'https://example.com';
+const baseURL = "https://example.com";
 
-
-test('urlPattern', t => {
-  let pattern = new URLPattern({baseURL, pathname: '/product/*?'});
-  t.true(pattern.test(baseURL + '/product/a/b'));
+test("urlPattern", (t) => {
+  let pattern = new URLPattern({ baseURL, pathname: "/product/*?" });
+  t.true(pattern.test(baseURL + "/product/a/b"));
 });
 
-test('export of urlPattern there?', async t => {
+test("export of urlPattern there?", async (t) => {
   /** overwrite global wil local version imported, so we know the export is in place */
-  const {URLPattern} = await import("../dist/index.js");
-  let pattern = new URLPattern({baseURL, pathname: '/product/*?'});
-  t.true(pattern.test(baseURL + '/product/a/b'));
-})
+  const { URLPattern } = await import("urlpattern-polyfill");
+  let pattern = new URLPattern({ baseURL, pathname: "/product/*?" });
+  t.true(pattern.test(baseURL + "/product/a/b"));
+});
