+++ test/can-load-as-ts.ts
@@ -1,14 +1,14 @@
-require( "../dist/index.cjs");
-const test = require( "ava"); 
+require("urlpattern-polyfill");
+const test = require("ava");
 
-const baseURL = 'https://example.com';
+const baseURL = "https://example.com";
 
-test('urlPattern', t => {
-  let pattern = new URLPattern({baseURL, pathname: '/product/*?'});
-  t.true(pattern.test(baseURL + '/product/a/b'));
+test("urlPattern", (t) => {
+  let pattern = new URLPattern({ baseURL, pathname: "/product/*?" });
+  t.true(pattern.test(baseURL + "/product/a/b"));
 });
 
-test('exports urlPattern', t => {
-  const { URLPattern } = require('../dist/index.cjs');
-  t.true(typeof URLPattern === 'function');
-})
+test("exports urlPattern", (t) => {
+  const { URLPattern } = require("urlpattern-polyfill");
+  t.true(typeof URLPattern === "function");
+});
