+++ test/index.js
@@ -1,15 +1,15 @@
-import test from 'ava';
-import { readFileSync } from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
+import test from "ava";
+import { readFileSync } from "fs";
+import path from "path";
+import { fileURLToPath } from "url";
 
 import { runTests } from "./wpt/wpt-test-runner.js";
 
-import '../dist/index.js';
+import "urlpattern-polyfill";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-const baseURL = 'https://example.com';
+const baseURL = "https://example.com";
 
 /*
 test('shorthands 1', t => {
@@ -84,7 +84,7 @@ test('JavaScript URL routing 2/2', t => {
 });
 */
 
-let rawdata = readFileSync(path.resolve(__dirname, 'urlpatterntestdata.json'));
+let rawdata = readFileSync(path.resolve(__dirname, "urlpatterntestdata.json"));
 let data = JSON.parse(rawdata);
 
 runTests(data, test);
